[Session] Deprecate session configurators in favor of event listeners - related to #12241
