+++ bundles/AdminBundle/Resources/config/services.yaml
@@ -4,16 +4,6 @@ services:
         autoconfigure: true
         public: false
 
-    #
-    # SESSION
-    #
-
-    Pimcore\Bundle\AdminBundle\Session\AdminSessionBagConfigurator:
-        arguments:
-            - '%pimcore.admin.session.attribute_bags%'
-        tags:
-            - { name: pimcore.session.configurator }
-
     Pimcore\Bundle\AdminBundle\Session\Handler\AdminSessionHandler:
         public: true
         calls:
