+++ doc/Development_Documentation/19_Development_Tools_and_Details/35_Working_with_Sessions.md
@@ -3,41 +3,44 @@
 If you need sessions, please use the native session handling provided by Symfony (configured through the `framework.session` config). 
 For details see [sessions docs](https://symfony.com/doc/5.2/components/http_foundation/sessions.html). 
 
-Pimcore adds the possibility to configure sessions before they are started through `SessionConfiguratorInterface` registered 
-as service with the `pimcore.session.configurator` tag. This is useful when you need a custom session bag for your bundle
-or application. 
+In case  you need to add a custom session bag for your bundle or application, then implement an EventListener to register the session bag before the session started.
 
-#### Sample Session Configurator
+#### Register a Session Bag through EventListener
 ```php
 <?php
  
-namespace TestBundle\Session\Configurator;
+namespace TestBundle\EventListener;
  
 use Pimcore\Session\SessionConfiguratorInterface;
 use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBag;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
  
-class SessionCartConfigurator implements SessionConfiguratorInterface
+class SessionBagListener implements EventSubscriberInterface
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function configure(SessionInterface $session)
+    public static function getSubscribedEvents()
     {
+        return [
+            //run after Symfony\Component\HttpKernel\EventListener\SessionListener
+            KernelEvents::REQUEST => ['onKernelRequest', 127],
+        ];
+    }
+    
+    /**
+     * @param RequestEvent $event
+     */
+    public function onKernelRequest(RequestEvent $event)
+    {
+        if (!$event->isMainRequest()) {
+            return;
+        }
+
+        $session = $event->getRequest()->getSession();
         $bag = new AttributeBag('_session_cart');
         $bag->setName('session_cart');
  
         $session->registerBag($bag);
     }
 }
-```
-
-#### Session Configurator Service Definition
-```yml
-services:
-    TestBundle\Session\Configurator\SessionCartConfigurator:
-        tags:
-            - { name: pimcore.session.configurator }
 ```
 
 #### Usage of Configured Session, e.g. in Controller
