+++ lib/Session/SessionConfigurator.php
@@ -18,6 +18,8 @@ namespace Pimcore\Session;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
 /**
+ * @deprecated will be removed in Pimcore 11
+ *
  * Handles a collection of session configurators.
  */
 class SessionConfigurator implements SessionConfiguratorInterface
