+++ tests/_support/Test/EcommerceTestCase.php
@@ -19,7 +19,7 @@ namespace Pimcore\Tests\Test;
 
 use Pimcore\Bundle\EcommerceFrameworkBundle\Environment;
 use Pimcore\Bundle\EcommerceFrameworkBundle\EnvironmentInterface;
-use Pimcore\Bundle\EcommerceFrameworkBundle\Tools\SessionConfigurator;
+use Pimcore\Bundle\EcommerceFrameworkBundle\EventListener\SessionBagListener;
 use Pimcore\Localization\LocaleService;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Session\Session;
@@ -52,10 +52,10 @@ abstract class EcommerceTestCase extends TestCase
         if (null === $this->session) {
             $this->session = new Session(new MockArraySessionStorage());
 
-            $configurator = new SessionConfigurator();
+            $configurator = new SessionBagListener();
             $configurator->configure($this->session);
 
-            $this->session->getBag(SessionConfigurator::ATTRIBUTE_BAG_CART)->set('carts', []);
+            $this->session->getBag(SessionBagListener::ATTRIBUTE_BAG_CART)->set('carts', []);
 
             $requestStack = \Pimcore::getContainer()->get('request_stack');
             if (!$request = $requestStack->getCurrentRequest()) {
